requirements.txt
discord.py
requests
python-dateutil
beautifulsoup4
aiohttp
pyyaml